add spring security auth

+++ src/main/java/com/graduate/model/User.java
@@ -36,7 +36,7 @@ public class User {
         this.id = id;
     }
 
-    public short getIsModerator() {
+    public byte getIsModerator() {
         return isModerator;
     }
